Separate code into components and clean things up generally.
